+++ src/main/java/kz/bars/familybudget/repository/RoleRepo.java
@@ -5,9 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigInteger;
+
 @Repository
 @Transactional
-public interface RoleRepo extends JpaRepository<Role, Long> {
+public interface RoleRepo extends JpaRepository<Role, BigInteger> {
 
     Role findByRole(String role);
